start work on unsaved changes
URGENT! Topic: Unsaved Changes

MIKO HERE. Working on the process regarding unsaved changes especially during opening new file and exiting application

Work has been done in the event of exiting application when there are unsaved changes. HOWEVER, THERE IS A BUG! Due to setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE), the application only exits when 'YES' is chosen from confirmation dialogue. The problem is that the confirmation dialogue appears only when there is/are unsaved change/s.

In the case of a new file with no changes/updates, the confirmation dialogue does not appear, hence, there is no way to exit application.

Currently finding ways to solve the problem.

In the case of unsaved changes in the event of opening file, no work has been done yet.